Let docker-compose read the password for postgres from .env file
Without this password, docker compose would fail to launch the database,
as the POSTGRES_PASSWORD is a required env var for this container/image.

This commit adds an .env template file, ensures the .env is ignored and
therefore not accidentally committed. And adds instructions to the
README.

The error from postgres is:

Error: Database is uninitialized and superuser password is not specified.
       You must specify POSTGRES_PASSWORD to a non-empty value for the
       superuser. For example, "-e POSTGRES_PASSWORD=password" on "docker run".

       You may also use "POSTGRES_HOST_AUTH_METHOD=trust" to allow all
       connections without a password. This is *not* recommended.

       See PostgreSQL documentation about "trust":
       https://www.postgresql.org/docs/current/auth-trust.html